requirements.txt
praw
pandas
boto3
python-dotenv